requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
transformers
accelerate
torch
